Update docker-entrypoint.sh
Added a check to make sure database.db is still a sane file to process. If the file appears to be corrupt (e.g. not JSON) due to some issue with writing it out, or a crash of z2m, it will move the corrupt database.db out of the way and create an empty database.db to allow z2m to startup.